[ZenTao] Task 4437 (#24)
* fix: drive useStateUnlessEmpty by ConfigValue, not PlanValue

The Optional+Computed `po` attribute hit "Provider produced inconsistent
result after apply: .po: was cty.StringVal(\"\"), but now
cty.StringVal(\"admin\")" whenever a user removed `po` from a previously-set
config. The existing modifier only acted when `req.PlanValue.IsUnknown()`,
but in this scenario the framework hands us a non-Unknown PlanValue (null or
"") — depending on Terraform Core / plugin-framework version — so the guard
returned early, the empty value flowed into Update, the server backfilled
"admin", and the post-apply state mismatched the planned value.

Switch the modifier to gate on `ConfigValue` (the unambiguous "did the user
set this?" signal) and explicitly set PlanValue in both branches: pin prior
state when it carries a real value, else set Unknown so the server-assigned
default wins.

Expand the schema test to drive the full grid — `(ConfigValue ∈ {null,
set}) × (StateValue ∈ {empty, non-empty}) × (PlanValue ∈ {Unknown, null,
""})` — locking the previously-broken `config-null + plan-empty` and
`config-null + plan-null` paths alongside the original Unknown path. Update
SKILL §6b-bis to record the new gating contract and the wider failure mode
so future schema additions audit against the right rule.

* Change Product to use controller_transport.

* Finish development.

* Finalize Program module codes.

* Standardize SKILL.md based on program.go implementation.

* Resolve test error.

+++ zentaoAPI/project_test.go
@@ -181,7 +181,7 @@ func TestCreateProject_BodyShape(t *testing.T) {
 		Begin:         "2026-05-09",
 		End:           "2026-12-31",
 		Parent:        3,
-		Products:      []int{1, 2},
+		Products:      []int64{1, 2},
 		WorkflowGroup: 2,
 		ACL:           "private",
 		PM:            "pm",
